change gettAllUsers path
