--- src/app/api/usersActions/getAllUsers.js
@@ -1,16 +1,0 @@
-import { getErrorMessage } from '@/utils/utils';
-import axios from 'axios';
-
-export const getAllUsers = async () => {
-  try {
-    const { data } = await axios(
-      `http://localhost:3000/users/`
-      // `${process.env.NEXT_PUBLIC_BACKEND_URL}/users`
-    );
-    return { data };
-  } catch (error) {
-    return {
-      error: getErrorMessage(error),
-    };
-  }
-};
